% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DockerRunApplication.R
\name{docker.f4f.run}
\alias{docker.f4f.run}
\title{Run docker container}
\usage{
docker.f4f.run()
}
\description{
This is the function to run the application through its docker image
}
\examples{
\dontrun{
    #running
     docker.f4f.run()

}
}
\author{
Pernice Simone
}
