% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F4Frun.R
\name{F4F.run}
\alias{F4F.run}
\title{Run F4F}
\usage{
F4F.run()
}
\description{
function to lunch the F4F shiny application.
}
\examples{
\dontrun{
F4F.run()
}
}
\author{
Pernice Simone, Baccega Daniele, Terrone Irene, Frattarola Marco.
}
