% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FORGE4FLAMErun.R
\name{FORGE4FLAME.run}
\alias{FORGE4FLAME.run}
\title{Run FORGE4FLAME}
\usage{
FORGE4FLAME.run(FromDocker = F)
}
\description{
Function to lunch the FORGE4FLAME shiny application.
}
\examples{
\dontrun{
FORGE4FLAME.run()
}
}
\author{
Pernice Simone, Baccega Daniele, Terrone Irene, Frattarola Marco.
}
